Move all database setup code into the database file
This will allow us to reuse the database setup in other parts of our
application, like a web app for the API.
