+++ scraper.rb
@@ -1,7 +1,6 @@
 #!/usr/bin/env ruby
 
 require 'capybara/poltergeist'
-require 'active_record'
 require_relative 'database_config.rb'
 
 def format_location_name_for_table(table_elm)
@@ -12,27 +11,6 @@ end
 
 def tableize(string)
   string.downcase.gsub('- ', '').gsub(/(\.| )/,'_')
-end
-
-class AqmRecord < ActiveRecord::Base; end
-
-unless AqmRecord.table_exists?
-  ActiveRecord::Schema.define do
-    create_table :aqm_records do |t|
-      t.string :location_name
-      t.string :scraped_at
-      t.string :latest_reading_recorded_at
-      t.string :pm2_5_concentration
-      t.string :pm10_concentration
-      t.string :co_concentration
-      t.string :no2_concentration
-      t.string :differential_temperature_lower
-      t.string :differential_temperature_upper
-      t.string :wind_speed
-      t.string :wind_direction
-      t.string :sigma
-    end
-  end
 end
 
 capybara = Capybara::Session.new(:poltergeist)
